[AMP-461] initial commit of OpenCL 2.0 C++AMP runtime
With the new architecture, it's now rather easy to add new C++AMP runtime
implementations.  In this commit we introduce a new OpenCL 2.0 C++AMP
runtime.  Right now it looks exactly the same as OpenCL 1.2, and in the
further commits we will gradually add SVM-codes cherry-picked from ocl20-exp
branch.